Moves details.tsx out of component folder
